+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @weaverse/react
 
+## 4.2.1
+
+### Patch Changes
+
+- @weaverse/core@4.2.1
+
+## 4.2.0
+
+### Patch Changes
+
+- @weaverse/core@4.2.0
+
 ## 4.1.2
 
 ### Patch Changes