+++ packages/shopify/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @weaverse/shopify
 
+## 4.2.1
+
+### Patch Changes
+
+- @weaverse/react@4.2.1
+
+## 4.2.0
+
+### Patch Changes
+
+- @weaverse/react@4.2.0
+
 ## 4.1.2
 
 ### Patch Changes